+++ app/(main)/(organization)/organization-profile.tsx
@@ -1,10 +1,10 @@
 import ImagePicker from '@/components/ui/image-picker/ImagePicker';
 import InputField from '@/components/ui/input/InputField';
-import Colors from '@/constants/Colors';
 import useBreakpoints from '@/hooks/use-breakpoints';
 import AppLayout from '@/layouts/app-layout';
+import styles from '@/styles/organization/OrganizationProfile.styles';
 import React, { useState } from 'react';
-import { View, Text, StyleSheet, Button, ScrollView } from 'react-native';
+import { View, Text, Button, ScrollView } from 'react-native';
 
 const OrganizationProfile: React.FC = () => {
   const organization = {
@@ -128,38 +128,6 @@ const OrganizationProfile: React.FC = () => {
   );
 };
 
-const styles = StyleSheet.create({
-  scrollViewContent: {
-    flex: 1,
-    alignItems: 'center',
-  },
-  container: {
-    padding: 20,
-    paddingTop: 0,
-    gap: 20,
-    justifyContent: 'space-between',
-    backgroundColor: Colors.regular.white,
-    width: '100%',
-  },
-  row: {
-    gap: 20,
-  },
-  image: {
-    borderRadius: 75,
-    width: 100,
-    height: 100,
-    borderColor: Colors.regular.primary,
-    borderWidth: 5,
-  },
-  imagePickerContainer: {
-    alignItems: 'center',
-  },
-  title: {
-    fontSize: 18,
-    fontWeight: '600',
-    textAlign: 'center',
-    paddingBottom: 20,
-  },
-});
+
 
 export default OrganizationProfile;
